refactor(ios): VoiceStudioViewModel(1051) errorCode 매핑 extension 분리 → 933줄
순수 함수(mapVoiceError/localizedVoiceMessage/extractServerErrorCode)와 String
확장을 VoiceStudioViewModel+ErrorMapping.swift 로 이동. 인스턴스 상태 미접근이라 안전.
